+++ apps/ui-pre/src/components/forms/SignInForm.tsx
@@ -4,9 +4,9 @@ import { Box, Stack, Button, useColorModeValue, Text } from '@chakra-ui/react';
 import { Form, Formik, FormikConfig } from 'formik';
 import { EmailInput, PasswordInput } from '@/components/inputs';
 import Link from '@/components/link/Link';
-import { FaGoogle } from 'react-icons/fa';
+import GoogleButton from '../buttons/GoogleButton';
 // utils
-import { GoogleAuthProvider, signInWithEmailAndPassword, signInWithPopup } from 'firebase/auth';
+import { signInWithEmailAndPassword } from 'firebase/auth';
 import * as Yup from 'yup';
 // constants
 import * as C from '@/constants';
@@ -16,36 +16,11 @@ import { useNavigate } from 'react-router-dom';
 import useFormatMessage from '@/hooks/useFormatMessage';
 import useAppToast from '@/hooks/useAppToast';
 
-const provider = new GoogleAuthProvider();
-
 type SignInFormConfig = FormikConfig<{ email: string; password: string }>;
 
 const SignInFormComponent: SignInFormConfig['component'] = ({ isSubmitting }) => {
   const t = useFormatMessage();
-  const auth = useAuth();
   const toast = useAppToast();
-  const navigate = useNavigate();
-
-  const handleGoogleSubmit = async () => {
-    try {
-      const {
-        user: { displayName },
-      } = await signInWithPopup(auth, provider);
-      toast({
-        description: t('signIn.form.toast.success.description', { displayName }),
-        status: 'success',
-        title: t('signIn.form.toast.success.title'),
-      });
-      navigate(C.ROUTES.DASHBOARD);
-    } catch (e) {
-      console.error(e);
-      toast({
-        description: t('signIn.form.toast.error.description'),
-        status: 'error',
-        title: t('signIn.form.toast.error.title'),
-      }); // TODO: catch all error codes
-    }
-  };
 
   return (
     <Form id='sign-in-form'>
@@ -62,15 +37,25 @@ const SignInFormComponent: SignInFormConfig['component'] = ({ isSubmitting }) =>
           <Button color='white' colorScheme='green' isLoading={isSubmitting} type='submit'>
             {t('signIn.form.submitButton.label')}
           </Button>
-          <Button
-            color='white'
-            colorScheme='red'
-            leftIcon={<FaGoogle />}
-            onClick={handleGoogleSubmit}
-            type='button'
-          >
-            {t('signIn.form.googleButton.label')}
-          </Button>
+          <GoogleButton
+            label={t('signIn.form.googleButton.label')}
+            onError={() => {
+              toast({
+                description: t('signIn.form.toast.error.description'),
+                status: 'error',
+                title: t('signIn.form.toast.error.title'),
+              }); // TODO: catch all error codes
+            }}
+            onSuccess={(result) => {
+              toast({
+                description: t('signIn.form.toast.success.description', {
+                  displayName: result.user.displayName,
+                }),
+                status: 'success',
+                title: t('signIn.form.toast.success.title'),
+              });
+            }}
+          />
           <Text align='center' id='sign-in-form-footer'>
             <FormattedMessage
               id='signIn.form.footer'
